Clean up some imports to higher level
