+++ mira/dkg/model.py
@@ -21,14 +21,10 @@ from fastapi.responses import FileResponse
 from pydantic import BaseModel, Field
 
 from mira.examples.sir import sir_bilayer, sir
-from mira.metamodel import NaturalConversion, Template, ControlledConversion
-from mira.metamodel.ops import stratify
+from mira.metamodel import NaturalConversion, Template, ControlledConversion, \
+    stratify, Concept, ModelComparisonGraphdata, TemplateModelDelta, \
+    TemplateModel, Parameter, simplify_rate_laws, aggregate_parameters
 from mira.modeling import Model
-from mira.metamodel.templates import Concept
-from mira.metamodel.comparison import ModelComparisonGraphdata, \
-    TemplateModelDelta
-from mira.metamodel.template_model import TemplateModel, Parameter
-from mira.metamodel.ops import simplify_rate_laws, aggregate_parameters
 from mira.modeling.bilayer import BilayerModel
 from mira.modeling.petri import PetriNetModel, PetriNetResponse
 from mira.modeling.viz import GraphicalModel
