+++ mira/examples/chime.py
@@ -1,8 +1,7 @@
 """CHIME SVIIvR."""
 
-from mira.metamodel import NaturalConversion
-from mira.metamodel.templates import ControlledConversion
-from mira.metamodel.template_model import TemplateModel
+from mira.metamodel import NaturalConversion, ControlledConversion, \
+    TemplateModel
 from .concepts import susceptible, infected, recovered
 
 infection = ControlledConversion(
