+++ mira/examples/jin2022.py
@@ -6,9 +6,8 @@
     than the simple flow in the graphical part of the model.
 """
 
-from mira.metamodel import NaturalConversion
-from mira.metamodel.templates import ControlledConversion
-from mira.metamodel.template_model import TemplateModel
+from mira.metamodel import NaturalConversion, ControlledConversion, \
+    TemplateModel
 from.concepts import susceptible, infected, recovered, exposed, dead
 
 __all__ = [
