+++ mira/modeling/petri.py
@@ -14,7 +14,7 @@ from typing import List, Optional
 from pydantic import BaseModel, Field
 from sympy.printing.mathml import mathml
 
-from mira.modeling import Model
+from . import Model
 
 
 class State(BaseModel):
