+++ mira/sources/petri.py
@@ -3,7 +3,6 @@ import ast
 import json
 from collections import defaultdict
 from mira.metamodel import *
-from mira.metamodel.template_model import TemplateModel, Initial, Parameter
 
 
 def template_model_from_petri_json_file(petri_json_file) -> TemplateModel:
