+++ tests/test_template.py
@@ -2,9 +2,8 @@ import json
 import sympy
 from mira.metamodel import ControlledConversion, Concept, NaturalConversion, \
     NaturalDegradation, Template, GroupedControlledConversion, \
-    GroupedControlledProduction
+    GroupedControlledProduction, TemplateModel
 from mira.metamodel.templates import Config
-from mira.metamodel.template_model import TemplateModel
 from mira.dkg.web_client import is_ontological_child_web
 
 # Provide to tests that are not meant to test ontological refinements;
